move button to separate file to fix docs build
